nodeless - lift file dialog service to /browser (#70541)
