+++ src/tsconfig.strictNullChecks.json
@@ -299,8 +299,9 @@
 		"./vs/workbench/services/decorations/browser/decorations.ts",
 		"./vs/workbench/services/decorations/browser/decorationsService.ts",
 		"./vs/workbench/services/decorations/test/browser/decorationsService.test.ts",
+		"./vs/workbench/services/dialogs/browser/remoteFileDialog.ts",
+		"./vs/workbench/services/dialogs/browser/fileDialogService.ts",
 		"./vs/workbench/services/dialogs/electron-browser/dialogService.ts",
-		"./vs/workbench/services/dialogs/electron-browser/remoteFileDialog.ts",
 		"./vs/workbench/services/editor/browser/codeEditorService.ts",
 		"./vs/workbench/services/editor/browser/editorService.ts",
 		"./vs/workbench/services/editor/common/editorGroupsService.ts",
