+++ src/vs/workbench/services/dialogs/browser/remoteFileDialog.ts
@@ -6,7 +6,6 @@
 import * as nls from 'vs/nls';
 import * as resources from 'vs/base/common/resources';
 import * as objects from 'vs/base/common/objects';
-import { RemoteFileService } from 'vs/workbench/services/files/node/remoteFileService';
 import { IFileService, IFileStat, FileKind } from 'vs/platform/files/common/files';
 import { IQuickInputService, IQuickPickItem, IQuickPick, IQuickInputButton } from 'vs/platform/quickinput/common/quickInput';
 import { URI } from 'vs/base/common/uri';
@@ -48,7 +47,7 @@ export class RemoteFileDialog {
 	private shouldOverwriteFile: boolean = false;
 
 	constructor(
-		@IFileService private readonly remoteFileService: RemoteFileService,
+		@IFileService private readonly fileService: IFileService,
 		@IQuickInputService private readonly quickInputService: IQuickInputService,
 		@IWindowService private readonly windowService: IWindowService,
 		@ILabelService private readonly labelService: ILabelService,
@@ -77,7 +76,7 @@ export class RemoteFileDialog {
 
 		return this.pickResource().then(async fileFolderUri => {
 			if (fileFolderUri) {
-				const stat = await this.remoteFileService.resolveFile(fileFolderUri);
+				const stat = await this.fileService.resolveFile(fileFolderUri);
 				return <IURIToOpen[]>[{ uri: fileFolderUri, typeHint: stat.isDirectory ? 'folder' : 'file' }];
 
 			}
@@ -106,7 +105,7 @@ export class RemoteFileDialog {
 
 	private getOptions(options: ISaveDialogOptions | IOpenDialogOptions): IOpenDialogOptions | undefined {
 		const defaultUri = options.defaultUri ? options.defaultUri : URI.from({ scheme: this.scheme, authority: this.remoteAuthority, path: '/' });
-		if ((this.scheme !== Schemas.file) && !this.remoteFileService.canHandleResource(defaultUri)) {
+		if ((this.scheme !== Schemas.file) && !this.fileService.canHandleResource(defaultUri)) {
 			this.notificationService.info(nls.localize('remoteFileDialog.notConnectedToRemote', 'File system provider for {0} is not available.', defaultUri.toString()));
 			return undefined;
 		}
@@ -141,7 +140,7 @@ export class RemoteFileDialog {
 		let ext: string = resources.extname(homedir);
 		if (this.options.defaultUri) {
 			try {
-				stat = await this.remoteFileService.resolveFile(this.options.defaultUri);
+				stat = await this.fileService.resolveFile(this.options.defaultUri);
 			} catch (e) {
 				// The file or folder doesn't exist
 			}
@@ -267,8 +266,8 @@ export class RemoteFileDialog {
 		let stat: IFileStat | undefined;
 		let statDirname: IFileStat | undefined;
 		try {
-			statDirname = await this.remoteFileService.resolveFile(inputUriDirname);
-			stat = await this.remoteFileService.resolveFile(inputUri);
+			statDirname = await this.fileService.resolveFile(inputUriDirname);
+			stat = await this.fileService.resolveFile(inputUri);
 		} catch (e) {
 			// do nothing
 		}
@@ -310,7 +309,7 @@ export class RemoteFileDialog {
 		if (this.endsWithSlash(value) || (!resources.isEqual(this.currentFolder, resources.dirname(valueUri), true) && resources.isEqualOrParent(this.currentFolder, resources.dirname(valueUri), true))) {
 			let stat: IFileStat | undefined;
 			try {
-				stat = await this.remoteFileService.resolveFile(valueUri);
+				stat = await this.fileService.resolveFile(valueUri);
 			} catch (e) {
 				// do nothing
 			}
@@ -321,7 +320,7 @@ export class RemoteFileDialog {
 				if (!resources.isEqual(this.currentFolder, inputUriDirname, true)) {
 					let statWithoutTrailing: IFileStat | undefined;
 					try {
-						statWithoutTrailing = await this.remoteFileService.resolveFile(inputUriDirname);
+						statWithoutTrailing = await this.fileService.resolveFile(inputUriDirname);
 					} catch (e) {
 						// do nothing
 					}
@@ -358,8 +357,8 @@ export class RemoteFileDialog {
 		let stat: IFileStat | undefined;
 		let statDirname: IFileStat | undefined;
 		try {
-			statDirname = await this.remoteFileService.resolveFile(resources.dirname(uri));
-			stat = await this.remoteFileService.resolveFile(uri);
+			statDirname = await this.fileService.resolveFile(resources.dirname(uri));
+			stat = await this.fileService.resolveFile(uri);
 		} catch (e) {
 			// do nothing
 		}
@@ -480,7 +479,7 @@ export class RemoteFileDialog {
 
 		const backDir = this.createBackItem(currentFolder);
 		try {
-			const fileNames = await this.remoteFileService.readFolder(currentFolder);
+			const fileNames = await this.fileService.readFolder(currentFolder);
 			const items = await Promise.all(fileNames.map(fileName => this.createItem(fileName, currentFolder)));
 			for (let item of items) {
 				if (item) {
@@ -528,7 +527,7 @@ export class RemoteFileDialog {
 	private async createItem(filename: string, parent: URI): Promise<FileQuickPickItem | undefined> {
 		let fullPath = resources.joinPath(parent, filename);
 		try {
-			const stat = await this.remoteFileService.resolveFile(fullPath);
+			const stat = await this.fileService.resolveFile(fullPath);
 			if (stat.isDirectory) {
 				filename = this.basenameWithTrailingSlash(fullPath);
 				return { label: filename, uri: fullPath, isFolder: true, iconClasses: getIconClasses(this.modelService, this.modeService, fullPath || undefined, FileKind.FOLDER) };
@@ -543,8 +542,8 @@ export class RemoteFileDialog {
 
 	private getDialogIcons(name: string): { light: URI, dark: URI } {
 		return {
-			dark: URI.parse(require.toUrl(`vs/workbench/services/dialogs/electron-browser/media/dark/${name}.svg`)),
-			light: URI.parse(require.toUrl(`vs/workbench/services/dialogs/electron-browser/media/light/${name}.svg`))
+			dark: URI.parse(require.toUrl(`vs/workbench/services/dialogs/browser/media/dark/${name}.svg`)),
+			light: URI.parse(require.toUrl(`vs/workbench/services/dialogs/browser/media/light/${name}.svg`))
 		};
 	}
 }
